celeryLatex/gui/requirements.txt
# model base #
numpy
pillow
onnxruntime
tokenizers
pyside6
keyboard
colorlog
# build #
# nuitka
# orderedset
